+++ src/test/test-schemas/complex-as-schema-schema.ts
@@ -1,0 +1,10 @@
+import Complex from '../../complex'
+
+const schema = new Complex({
+  field: String,
+  field2: Number
+})
+
+export const typescriptSchemas: any[] = [
+  {name: 'CustomName', schema}
+]