+++ test-output/api-test/no-extra-exports-schema.schematar.ts
@@ -1,7 +1,0 @@
-// Generated file, do not edit!
-
-// tslint:disable array-type
-// @ts-ignore -- ignore possibly unused type parameters
-export interface NoExtraExportsSchemaBase<IDType, DateType> {
-  field: string
-}